Use solarized colors for ls output
There may be a better way to do this, but I've aliased ls to gls,
because gls (the GNU version of ls), has better ability to customize
color output. This requires installing coreutils, and while it is
possible to add the GNU versions to the PATH, I'm not yet sold on doing
so globally.